// src/interface/price-type.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface Price {
  success: boolean;
  data: any[] | DataClass;
}

export interface DataClass {
  price_overview: PriceOverview;
}

export interface PriceOverview {
  currency: string;
  initial: number;
  final: number;
  discount_percent: number;
  initial_formatted: string;
  final_formatted: string;
}