// src/interface/app-type.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface App {
  response: Response;
}

export interface Response {
  game_count: number;
  games: Game[];
}

export interface Game {
  appid: number;
  name: string;
  playtime_forever: number;
  img_icon_url: string;
  img_logo_url: string;
  playtime_windows_forever: number;
  playtime_mac_forever: number;
  playtime_linux_forever: number;
  has_community_visible_stats?: boolean;
  playtime_2weeks?: number;
  discount_percent?: number;
}